refactor(fun.roasts): rewrite
